Refactor components in FlightPage
